Added RelaxingSound Class and SleepEnviormentClass. Fixed bugs.
Added new classes to the calm down feature, started working on the Sleep Environment class and made changes to main and AlarmPanel.
Fixed bug in AlarmPanel and method playAlarmSound.
(Second attempt to structure better)
